requirements.txt
Django==2.2.6
djangorestframework==3.10.3
pytz==2019.3
sqlparse==0.3.0
gunicorn==19.9.0
psycopg2-binary==2.7.6.1
